requirements.txt
huggingface_hub>=0.20.3
sentence-transformers>=2.2.0
anthropic>=0.40.0
openai>=1.0.0
chromadb>=0.4.0,<0.6.0
faiss-cpu>=1.7.4
mem0ai>=0.1.0
gradio>=4.44.0,<5.0.0
pandas>=2.0.0
matplotlib>=3.7.0
rich>=13.0.0
reportlab>=4.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytest>=7.4.0
